fix: move board fetching logic to api folder
