--- src/api/README.md
@@ -1,1 +1,0 @@
-##Api requests
