+++ src/pages/Board.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { useAppDispatch, useAppSelector } from '../hooks/redux';
 import { RootState } from '../store/store';
-import { fetchBoard } from '../store/reducers/actionCreators';
+import { fetchBoard } from '../api/board.api';
 import { useParams } from 'react-router-dom';
 
 const Board: React.FC = () => {
@@ -9,7 +9,10 @@ const Board: React.FC = () => {
   const { boards } = useAppSelector((state: RootState) => state);
   const dispatch = useAppDispatch();
   useEffect(() => {
-    dispatch(fetchBoard(params.id));
+    const boardId = params.id;
+    if (boardId) {
+      dispatch(fetchBoard(boardId));
+    }
   }, [dispatch, params.id]);
 
   return (
